Moved triggers to URLBar
